requirements.txt
Django==2.2.10
gunicorn==19.9.0
django-cors-headers==3.2.1
django-extensions==2.2.9
dj-rest-auth==1.0.4
djangorestframework==3.11.0
Pillow==7.1.2
django-filter==2.2.0

# Heroku staging debug tools
django-debug-toolbar==2.2
django-querycount==0.7.0

# Database
dj-database-url==0.5.0
psycopg2==2.8.5
Pillow==7.1.2
boto3==1.12.29

# Storage
whitenoise==3.3.0
django-storages==1.9.1
